+++ src/pages/sseung/Login/JoinWrap.js
@@ -1,0 +1,11 @@
+import { Link } from 'react-router-dom';
+
+const JoinWrap = () => {
+  return (
+    <div className="joinWrap">
+      계정이 없으신가요 ? <Link to="#">가입하기</Link>
+    </div>
+  );
+};
+
+export default JoinWrap;
