+++ src/pages/sseung/Main/Main.js
@@ -4,6 +4,7 @@ import Nav from '../../../components/Nav/Nav';
 import Feed from './Feed';
 import './Main.scss';
 import MainRight from './MainRight';
+import StoriesWrap from './StoriesWrap';
 
 const Main = () => {
   const feedData = [
@@ -94,59 +95,7 @@ const Main = () => {
       <main>
         <div>
           <section className="feeds">
-            <article className="stories_wrap">
-              <ul className="stories">
-                <li className="story">
-                  <div className="user_img">
-                    <div className="profile_img" />
-                  </div>
-                  <p>AAAAA</p>
-                </li>
-                <li className="story">
-                  <div className="user_img">
-                    <div className="profile_img" />
-                  </div>
-                  <p>BBBBBBB</p>
-                </li>
-                <li className="story">
-                  <div className="user_img">
-                    <div className="profile_img" />
-                  </div>
-                  <p>CCCCC</p>
-                </li>
-                <li className="story">
-                  <div className="user_img">
-                    <div className="profile_img" />
-                  </div>
-                  <p>DDDDDDDDD</p>
-                </li>
-                <li className="story">
-                  <div className="user_img">
-                    <div className="profile_img" />
-                  </div>
-                  <p>EEEE</p>
-                </li>
-                <li className="story">
-                  <div className="user_img">
-                    <div className="profile_img" />
-                  </div>
-                  <p>FFF</p>
-                </li>
-                <li className="story">
-                  <div className="user_img">
-                    <div className="profile_img" />
-                  </div>
-                  <p>GGGGGGGG</p>
-                </li>
-                <li className="story">
-                  <div className="user_img">
-                    <div className="profile_img" />
-                  </div>
-                  <p>mschoise</p>
-                </li>
-              </ul>
-            </article>
-
+            <StoriesWrap />
             {feedData.map(it => (
               <Feed key={it.id} data={it} />
             ))}
