+++ src/pages/sseung/Main/StoriesWrap.js
@@ -1,0 +1,58 @@
+const StoriesWrap = () => {
+  return (
+    <article className="stories_wrap">
+      <ul className="stories">
+        <li className="story">
+          <div className="user_img">
+            <div className="profile_img" />
+          </div>
+          <p>AAAAA</p>
+        </li>
+        <li className="story">
+          <div className="user_img">
+            <div className="profile_img" />
+          </div>
+          <p>BBBBBBB</p>
+        </li>
+        <li className="story">
+          <div className="user_img">
+            <div className="profile_img" />
+          </div>
+          <p>CCCCC</p>
+        </li>
+        <li className="story">
+          <div className="user_img">
+            <div className="profile_img" />
+          </div>
+          <p>DDDDDDDDD</p>
+        </li>
+        <li className="story">
+          <div className="user_img">
+            <div className="profile_img" />
+          </div>
+          <p>EEEE</p>
+        </li>
+        <li className="story">
+          <div className="user_img">
+            <div className="profile_img" />
+          </div>
+          <p>FFF</p>
+        </li>
+        <li className="story">
+          <div className="user_img">
+            <div className="profile_img" />
+          </div>
+          <p>GGGGGGGG</p>
+        </li>
+        <li className="story">
+          <div className="user_img">
+            <div className="profile_img" />
+          </div>
+          <p>mschoise</p>
+        </li>
+      </ul>
+    </article>
+  );
+};
+
+export default StoriesWrap;
